Refactor api to root
